+++ Engine/Source/Runtime/Core/Public/Containers/StaticBitArray.h
@@ -329,7 +329,7 @@ public:
 			const WordType Bits = Words[WordIndex];
 			ASSUME(Bits != 0);
 
-			const int32 LowestBitIndex = FMath::CountTrailingZeros64(Bits) + (WordIndex << NumBitsPerWordLog2);
+			const int32 LowestBitIndex = (int32)FMath::CountTrailingZeros64(Bits) + (WordIndex << NumBitsPerWordLog2);
 			if (LowestBitIndex < LocalNumBits)
 			{
 				return LowestBitIndex;
